Fix weather station hanging
This commit fixes an issue where the weather station hangs. The
issue was caused by the fact that the weather station established
a new HTTP session every time it queried the weather server,
and by the weather server taking longer and longer to allow a
new session. Now, the weather station persists the HTTP session.